+++ src/helpers/folders.ts
@@ -5,9 +5,11 @@
  */
 import { Account, FOLDERS } from '@zextras/carbonio-shell-ui';
 import { find } from 'lodash';
-import { getFolder } from '../carbonio-ui-commons/store/zustand/folder';
-import type { Folder, Folders, LinkFolder } from '../carbonio-ui-commons/types/folder';
+import { getFolder, useFolderStore } from '../carbonio-ui-commons/store/zustand/folder';
+import type { Folder, Folders } from '../carbonio-ui-commons/types/folder';
+import { Conversation } from '../types';
 import type { MailMessage } from '../types';
+import { isConversation } from './messages';
 
 /*
  * Describe the folder id syntax
@@ -213,4 +215,45 @@ export const isInboxSubfolder = ({
 	}
 
 	return path.toLowerCase().startsWith('/inbox');
+};
+
+/**
+ * Returns the parent folder id of the given message.
+ *
+ * In most of the cases the id coincide with the "parent" property of
+ * the message, but they differ when the message is contained in a linked folder: in this
+ * case the "parent" refers to the original owner folder zid and id, so there is the need
+ * to "translate" those ids
+ *
+ * @param item - Message or conversation
+ */
+export const getParentFolderId = (item: MailMessage | Conversation): string | null => {
+	if (!item) {
+		return null;
+	}
+
+	const parentId = isConversation(item) ? item.messages[0].parent : item.parent;
+	const parentParts = getFolderIdParts(parentId);
+
+	/*
+	 * If the zid is not present the id is referring to a normal user's folder
+	 * so there is no need to search among the stored folders
+	 */
+	if (!parentParts.zid) {
+		return parentId;
+	}
+
+	/*
+	 * If the parentId is in the form zid:(r)id it could refer to a shared account folder
+	 * or a linked folder.
+	 */
+
+	// First attempt: search for a shared account folder among the stored folders.
+	if (getFolder(parentId)) {
+		return parentId;
+	}
+
+	// Second attempt: search for an entry on the links id map, or return null otherwise
+	const state = useFolderStore.getState();
+	return state.linksIdMap[parentId] ?? null;
 };