+++ src/helpers/tests/folders.test.ts
@@ -4,17 +4,21 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 import { getUserAccount } from '@zextras/carbonio-shell-ui';
+import { FOLDER_VIEW } from '../../carbonio-ui-commons/constants';
 import { FOLDERS } from '../../carbonio-ui-commons/test/mocks/carbonio-shell-ui-constants';
 import { populateFoldersStore } from '../../carbonio-ui-commons/test/mocks/store/folders';
 import {
 	getFolder,
 	getFoldersArray,
+	getLinksArray,
 	getRootsMap
 } from '../../carbonio-ui-commons/store/zustand/folder/hooks';
 import { getMocksContext } from '../../carbonio-ui-commons/test/mocks/utils/mocks-context';
+import { generateMessage } from '../../tests/generators/generateMessage';
 import {
 	getFolderIdParts,
 	getFolderOwnerAccountName,
+	getParentFolderId,
 	isDraft,
 	isInbox,
 	isInboxSubfolder,
@@ -325,5 +329,62 @@ describe('isTrashed', () => {
 	test('The inbox folder (passed by id) is not recognized as trashed', () => {
 		populateFoldersStore();
 		expect(isTrashed({ folderId: FOLDERS.INBOX })).toBe(false);
+	});
+});
+
+describe('getParentFolderId', () => {
+	test('if the parameter is falsy null is returned', () => {
+		expect(
+			getParentFolderId(
+				// Testing the case with a falsy parameter
+				// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+				// @ts-ignore
+				null
+			)
+		).toBeNull();
+		expect(
+			getParentFolderId(
+				// Testing the case with a falsy parameter
+				// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+				// @ts-ignore
+				undefined
+			)
+		).toBeNull();
+		expect(
+			getParentFolderId(
+				// Testing the case with a falsy parameter
+				// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+				// @ts-ignore
+				false
+			)
+		).toBeNull();
+	});
+
+	test('if the parameter is a message with the parent set to 12345 then 12345 is returned', () => {
+		populateFoldersStore();
+		const msg = generateMessage({ folderId: '12345' });
+		expect(getParentFolderId(msg)).toBe('12345');
+	});
+
+	test('if the parameter is a message of a shared account folder the shared folder id is returned', () => {
+		populateFoldersStore();
+		const sharedAccountIdentity = getMocksContext().identities.sendAs[0].identity;
+		const sharedAccountInbox = getFolder(`${sharedAccountIdentity.id}:${FOLDERS.INBOX}`);
+		const msg = generateMessage({ folderId: sharedAccountInbox?.id });
+		expect(getParentFolderId(msg)).toBe(sharedAccountInbox?.id);
+	});
+
+	test('if the parameter is a message of a linked folder the folder id is returned', () => {
+		populateFoldersStore();
+		const links = getLinksArray(FOLDER_VIEW.message);
+		const linkFolder = links?.[0];
+		const msg = generateMessage({ folderId: linkFolder.id });
+		expect(getParentFolderId(msg)).toBe(linkFolder.id);
+	});
+
+	test('if the parameter is a message with a parent which is not a user account folder, shared account folder, linked folder then null is returned', () => {
+		populateFoldersStore();
+		const msg = generateMessage({ folderId: 'supercalifragilisticexpialidocious:42' });
+		expect(getParentFolderId(msg)).toBeNull();
 	});
 });